feat: 최근 학습 정보 조회 로직 수정(#22)
