+++ src/main/java/gravit/code/domain/chapter/domain/ChapterRepository.java
@@ -1,11 +1,6 @@
 package gravit.code.domain.chapter.domain;
 
-import gravit.code.domain.learning.dto.response.RecentLearningInfo;
-
-import java.util.Optional;
-
 public interface ChapterRepository {
-    Optional<RecentLearningInfo> findRecentLearningChapter(Long userId);
     Chapter save(Chapter chapter);
     Long getTotalUnitsByChapterId(Long chapterId);
 }
