+++ src/main/java/gravit/code/domain/mainPage/facade/MainPageFacade.java
@@ -1,8 +1,8 @@
 package gravit.code.domain.mainPage.facade;
 
-import gravit.code.domain.chapter.service.ChapterService;
-import gravit.code.domain.learning.dto.response.RecentLearningInfo;
 import gravit.code.domain.mainPage.dto.response.MainPageResponse;
+import gravit.code.domain.recentLearning.dto.response.RecentLearningInfo;
+import gravit.code.domain.recentLearning.service.RecentLearningService;
 import gravit.code.domain.user.dto.response.UserMainPageInfo;
 import gravit.code.domain.user.service.UserService;
 import gravit.code.domain.userLeague.service.UserLeagueService;
@@ -10,13 +10,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
-
 @Service
 @RequiredArgsConstructor
 public class MainPageFacade {
 
-    private final ChapterService chapterService;
+    private final RecentLearningService recentLearningService;
     private final UserService userService;
     private final UserLeagueService userLeagueService;
 
@@ -27,24 +25,15 @@ public class MainPageFacade {
 
         String league = userLeagueService.getUserLeagueName(userId);
 
-        Optional<RecentLearningInfo> recentLearningChapter = chapterService.getRecentLearningChapter(userId);
+        RecentLearningInfo recentLearningInfo = recentLearningService.getRecentLearningInfo(userId);
 
-        if(recentLearningChapter.isPresent()) {
-            return MainPageResponse.create(
-                    userMainPageInfo.nickname(),
-                    userMainPageInfo.level(),
-                    userMainPageInfo.xp(),
-                    league,
-                    recentLearningChapter.get()
-            );
-        }else{
-            return MainPageResponse.create(
-                    userMainPageInfo.nickname(),
-                    userMainPageInfo.level(),
-                    userMainPageInfo.xp(),
-                    league,
-                    new RecentLearningInfo(-1L, "-", "-", -1L, -1L)
-            );
-        }
+
+        return MainPageResponse.create(
+                userMainPageInfo.nickname(),
+                userMainPageInfo.level(),
+                userMainPageInfo.xp(),
+                league,
+                recentLearningInfo
+        );
     }
 }
