+++ src/main/java/gravit/code/domain/recentLearning/domain/RecentLearningRepository.java
@@ -1,9 +1,12 @@
 package gravit.code.domain.recentLearning.domain;
 
+import gravit.code.domain.recentLearning.dto.response.RecentLearningInfo;
+
 import java.util.Optional;
 
 public interface RecentLearningRepository {
 
     RecentLearning save(RecentLearning recentLearning);
     Optional<RecentLearning> findByUserId(Long userId);
+    Optional<RecentLearningInfo> findRecentLearningInfoByUserId(Long userId);
 }
