+++ src/main/java/gravit/code/domain/recentLearning/infrastructure/RecentLearningRepositoryImpl.java
@@ -2,6 +2,7 @@ package gravit.code.domain.recentLearning.infrastructure;
 
 import gravit.code.domain.recentLearning.domain.RecentLearning;
 import gravit.code.domain.recentLearning.domain.RecentLearningRepository;
+import gravit.code.domain.recentLearning.dto.response.RecentLearningInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -21,5 +22,10 @@ public class RecentLearningRepositoryImpl implements RecentLearningRepository {
     @Override
     public Optional<RecentLearning> findByUserId(Long userId) {
         return recentLearningJpaRepository.findByUserId(userId);
+    }
+
+    @Override
+    public Optional<RecentLearningInfo> findRecentLearningInfoByUserId(Long userId) {
+        return recentLearningJpaRepository.findRecentLearningInfoByUserId(userId);
     }
 }
